requirements.txt
aiogram~=3.5.0
requests~=2.31.0
aiofiles~=23.2.1
aiohttp~=3.9.5
aiosmtplib~=3.0.1
whisper~=1.1.10
openai-whisper
openai~=0.28.0
googlemaps~=4.10.0
python-dotenv~=1.0.1
requests-oauthlib~=2.0.0
fuzzywuzzy~=0.18.0
python-Levenshtein~=0.26.0
beautifulsoup4~=4.12.3
google-auth~=2.35.0
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.147.0
